📬生成反向代理流量分析报告，包括RequestData、CacheData缓存、封禁IP、日志数据、爬虫数据、服务器状态等信息的报告，生成pdf于./Report/Re/ReverseReport.pdf

+++ global/global.go
@@ -22,10 +22,10 @@ type RequestData struct {
 	Times int64  `json:"times"`
 }
 type CacheData struct {
-	ReqTimes  int64 `json:"ReqTimes"`
-	HintTimes int64 `json:"HintTimes"`
-	HintRate  int64 `json:"HintRate"`
-	Memory    int64 `json:"memory"`
+	ReqTimes  int64   `json:"ReqTimes"`
+	HintTimes int64   `json:"HintTimes"`
+	HintRate  float64 `json:"HintRate"`
+	Memory    int64   `json:"memory"`
 }
 type LogsData struct {
 	InfoTimes  int64 `json:"InfoTimes"`
